+++ audiorep/domain/__init__.py
@@ -2,6 +2,7 @@
 from audiorep.domain.track import Track, AudioFormat, TrackSource
 from audiorep.domain.album import Album
 from audiorep.domain.artist import Artist
+from audiorep.domain.label import Label
 from audiorep.domain.playlist import Playlist, PlaylistEntry
 from audiorep.domain.cd_disc import CDDisc, CDTrack, RipStatus
 from audiorep.domain.radio_station import RadioStation
@@ -10,6 +11,7 @@ __all__ = [
     "Track", "AudioFormat", "TrackSource",
     "Album",
     "Artist",
+    "Label",
     "Playlist", "PlaylistEntry",
     "CDDisc", "CDTrack", "RipStatus",
     "RadioStation",